requirements.txt
asgiref==3.9.1
asttokens==3.0.0
decorator==5.2.1
Django==5.2.5
django-extensions==4.1
djangorestframework==3.16.1
executing==2.2.0
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pluggy==1.6.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
